\name{holidayDatabase}

\alias{holidayDatabase}


\alias{Septuagesima}
\alias{Quinquagesima}
\alias{AshWednesday}
\alias{PalmSunday}
\alias{GoodFriday}
\alias{Easter}
\alias{EasterSunday}
\alias{EasterMonday}
\alias{RogationSunday}
\alias{Ascension}
\alias{Pentecost}
\alias{PentecostMonday}
\alias{TrinitySunday}
\alias{CorpusChristi}
\alias{ChristTheKing}
\alias{Advent1st}
\alias{Advent2nd}
\alias{Advent3rd}
\alias{Advent4th}
\alias{ChristmasEve}
\alias{ChristmasDay}
\alias{BoxingDay}
\alias{NewYearsDay}
\alias{SolemnityOfMary}
\alias{Epiphany}
\alias{PresentationOfLord}
\alias{Annunciation}
\alias{TransfigurationOfLord}
\alias{AssumptionOfMary}
\alias{BirthOfVirginMary}
\alias{CelebrationOfHolyCross}
\alias{MassOfArchangels}
\alias{AllSaints}
\alias{AllSouls}
\alias{LaborDay}
\alias{CHBerchtoldsDay}
\alias{CHSechselaeuten}
\alias{CHAscension}
\alias{CHConfederationDay}
\alias{CHKnabenschiessen}
\alias{GBMayDay}
\alias{GBBankHoliday}
\alias{GBSummerBankHoliday}
\alias{GBMilleniumDay}
\alias{DEAscension}
\alias{DECorpusChristi}
\alias{DEGermanUnity}
\alias{DEChristmasEve}
\alias{DENewYearsEve}
\alias{FRFetDeLaVictoire1945}
\alias{FRAscension}
\alias{FRBastilleDay}
\alias{FRAssumptionVirginMary}
\alias{FRAllSaints}
\alias{FRArmisticeDay}
\alias{ITEpiphany}
\alias{ITLiberationDay}
\alias{ITAssumptionOfVirginMary}
\alias{ITAllSaints}
\alias{ITStAmrose}
\alias{ITImmaculateConception}
\alias{USDecorationMemorialDay}
\alias{USPresidentsDay}
\alias{USNewYearsDay}
\alias{USInaugurationDay}
\alias{USMLKingsBirthday}
\alias{USLincolnsBirthday}
\alias{USWashingtonsBirthday}
\alias{USMemorialDay}
\alias{USIndependenceDay}
\alias{USLaborDay}
\alias{USColumbusDay}
\alias{USElectionDay}
\alias{USVeteransDay}
\alias{USThanksgivingDay}
\alias{USChristmasDay}
\alias{USCPulaskisBirthday}
\alias{USGoodFriday}
\alias{CAVictoriaDay}
\alias{CACanadaDay}
\alias{CACivicProvincialHoliday}
\alias{CALabourDay}
\alias{CAThanksgivingDay}
\alias{CaRemembranceDay}
\alias{JPNewYearsDay}
\alias{JPGantan}
\alias{JPBankHolidayJan2}
\alias{JPBankHolidayJan3}
\alias{JPComingOfAgeDay}
\alias{JPSeijinNoHi}
\alias{JPNatFoundationDay}
\alias{JPKenkokuKinenNoHi}
\alias{JPGreeneryDay}
\alias{JPMidoriNoHi}
\alias{JPConstitutionDay}
\alias{JPKenpouKinenBi}
\alias{JPNationHoliday}
\alias{JPKokuminNoKyujitu}
\alias{JPChildrensDay}
\alias{JPKodomoNoHi}
\alias{JPMarineDay}
\alias{JPUmiNoHi}
\alias{JPRespectForTheAgedDay}
\alias{JPKeirouNOhi}
\alias{JPAutumnalEquinox}
\alias{JPShuubunNoHi}
\alias{JPHealthandSportsDay}
\alias{JPTaiikuNoHi}
\alias{JPNationalCultureDay}
\alias{JPBunkaNoHi}
\alias{JPThanksgivingDay}
\alias{JPKinrouKanshaNoHi}
\alias{JPEmperorsBirthday}
\alias{JPTennouTanjyouBi}
\alias{JPBankHolidayDec31}


\title{Holiday Calendars and Utilities}


\description{
    
    A collection of functions and methods dealing with holiday 
    calendars. 
   
}


\usage{
Septuagesima(year)
Quinquagesima(year)
AshWednesday(year)
PalmSunday(year)
GoodFriday(year)
Easter(year)
EasterSunday(year)
EasterMonday(year)
RogationSunday(year)
Ascension(year)
Pentecost(year)
PentecostMonday(year)
TrinitySunday(year)
CorpusChristi(year)
ChristTheKing(year)
Advent1st(year)
Advent2nd(year)
Advent3rd(year)
Advent4th(year)
ChristmasEve(year)
ChristmasDay(year)
BoxingDay(year)
NewYearsDay(year)
SolemnityOfMary(year)
Epiphany(year)
PresentationOfLord(year)
Annunciation(year)
TransfigurationOfLord(year)
AssumptionOfMary(year)
BirthOfVirginMary(year)
CelebrationOfHolyCross(year)
MassOfArchangels(year)
AllSaints(year)
AllSouls(year)
LaborDay(year)
CHBerchtoldsDay(year)
CHSechselaeuten(year)
CHAscension(year)
CHConfederationDay(year)
CHKnabenschiessen(year)
GBMayDay(year)
GBBankHoliday(year)
GBSummerBankHoliday(year)
GBMilleniumDay(year)
DEAscension(year)
DECorpusChristi(year)
DEGermanUnity(year)
DEChristmasEve(year)
DENewYearsEve(year)
FRFetDeLaVictoire1945(year)
FRAscension(year)
FRBastilleDay(year)
FRAssumptionVirginMary(year)
FRAllSaints(year)
FRArmisticeDay(year)
ITEpiphany(year)
ITLiberationDay(year)
ITAssumptionOfVirginMary(year)
ITAllSaints(year)
ITStAmrose(year)
ITImmaculateConception(year)
USDecorationMemorialDay(year)
USPresidentsDay(year)
USNewYearsDay(year)
USInaugurationDay(year)
USMLKingsBirthday(year)
USLincolnsBirthday(year)
USWashingtonsBirthday(year)
USMemorialDay(year)
USIndependenceDay(year)
USLaborDay(year)
USColumbusDay(year)
USElectionDay(year)
USVeteransDay(year)
USThanksgivingDay(year)
USChristmasDay(year)
USCPulaskisBirthday(year)
USGoodFriday(year)
CAVictoriaDay(year)
CACanadaDay(year)
CACivicProvincialHoliday(year)
CALabourDay(year)
CAThanksgivingDay(year)
CaRemembranceDay(year)
JPNewYearsDay(year)
JPGantan(year)
JPBankHolidayJan2(year)
JPBankHolidayJan3(year)
JPComingOfAgeDay(year)
JPSeijinNoHi(year)
JPNatFoundationDay(year)
JPKenkokuKinenNoHi(year)
JPGreeneryDay(year)
JPMidoriNoHi(year)
JPConstitutionDay(year)
JPKenpouKinenBi(year)
JPNationHoliday(year)
JPKokuminNoKyujitu(year)
JPChildrensDay(year)
JPKodomoNoHi(year)
JPMarineDay(year)
JPUmiNoHi(year)
JPRespectForTheAgedDay(year)
JPKeirouNOhi(year)
JPAutumnalEquinox(year)
JPShuubunNoHi(year)
JPHealthandSportsDay(year)
JPTaiikuNoHi(year)
JPNationalCultureDay(year)
JPBunkaNoHi(year)
JPThanksgivingDay(year)
JPKinrouKanshaNoHi(year)
JPEmperorsBirthday(year)
JPTennouTanjyouBi(year)
JPBankHolidayDec31(year)
}


\arguments{

    \item{year}{ 
        an integer value or vector of year numbers including the
        century. These are integers of the form \code{CCYY}, e.g. 
        \code{2000}.
        }
  
}


\value{

    Returns an ISO-8601 formatted calendar date \code{CCYYMMDD}, an
    object of class \code{"sdate"} which represents an integer
    value or vector.
   
}


\examples{
## no examples ...
}


\author{

    Diethelm Wuertz for this R-port.
    
}


\keyword{chron}

