\name{ghSlider}


\alias{ghSlider}


\title{Generalized Hyperbolic Distribution Slider}


\description{
    
    Displays interactively the dependence of the 
    generalized hyperbolic distribution on its parameters.
    
}


\usage{
ghSlider()
}


\value{

    a tcl/tk based graphical user interface.
    
    This is a nice display for educational purposes to investigate
    the densities and probabilities of the generalized hyperbolic 
    distribution.
}


\examples{   
## ghSlider -
   # ghSlider()
}


\keyword{distribution}

