\name{gldMode}
\alias{gldMode}


\title{Generalized Lambda Distribution Mode}

\description{

  Computes the mode of the generalized lambda distribution.

}

\usage{
gldMode(lambda1 = 0, lambda2 = -1, lambda3 = -1/8, lambda4 = -1/8)
}

\arguments{
  \item{lambda1}{
    location parameter.
  }
  \item{lambda2}{
    scale parameter.
  }
  \item{lambda3}{
    first shape parameter.
  }
  \item{lambda4}{
    second shape parameter.
  }
}

\author{
  Implemented by Diethelm Wuertz
}
