\name{hypSlider}


\alias{hypSlider}


\title{Hyperbolic Distribution Slider}


\description{
    
    Displays interactively the dependence of the 
    hyperbolic distribution on its parameters.
    
}


\usage{
hypSlider()
}


\value{

    a tcl/tk based graphical user interface.
    
    This is a nice display for educational purposes to investigate
    the densities and probabilities of the hyperbolic distribution.
}


\examples{   
## hypSlider -
   #
}


\keyword{distribution}

