% see Repeated descriptions of re-exported functions #408 at
%   https://github.com/klutometis/roxygen/issues/408
\docType{import}
\name{fBasics_reexports}
\alias{fBasics_reexports}
\alias{as.timeSeries}

\title{Objects exported from other packages}
\description{
  These objects are imported from other packages. Follow the links
  below to see their documentation.

  \describe{
    \item{lagged}{\code{\link[timeSeries]{as.timeSeries}}}
  }
}
\keyword{internal}
