\name{finCenter}
\alias{finCenter}
\alias{finCenter<-}

\title{Financial Center of to a timeSeries}

\description{
  Print or assign new financial center to an object.
}

\usage{
finCenter(x)
finCenter(x) <- value
}


\arguments{
  \item{x}{
    a \code{timeSeries} or \code{timeDate} object.
  }
  \item{value}{
    a character with the the location of the  financial center named 
    as "continent/city".
  }
}