\name{nigSlider}


\alias{nigSlider}


\title{nigerbolic Distribution Slider}


\description{
    
    Displays interactively the dependence of the 
    nigerbolic distribution on its parameters.
    
}


\usage{
nigSlider()
}


\value{

    a tcl/tk based graphical user interface.
    
    This is a nice display for educational purposes to investigate
    the densities and probabilities of the invetrse Gaussian 
    distribution.
    
}


\examples{   
## nigSlider -
   # nigSlider()
}


\keyword{distribution}

