\name{colorTable}
\alias{colorTable}


\title{Table of colors}

\description{
    
  Displays a table of color codes and plots the colors themselves.
        
}

\usage{ 
colorTable(cex = 0.7)
}

\arguments{
  
  \item{cex}{
    
    a numeric value, determines the character size in the color plot,
    the default is 0.7.
    
  }  
}

\value{
  a table of plot colors with the associated color numbers
}

\seealso{
    \code{\link{characterTable}},
    \code{\link{symbolTable}}
}

\examples{   
## Color Table:
   colorTable()
}

\keyword{programming}
