\name{tsHessian}

\alias{tsHessian}

\title{Two sided approximated Hessian}

\description{
  Computes two sided (TS) approximated Hessian.
}

\usage{
tsHessian(x, fun, ...)
}

\arguments{

  \item{x}{argument to be passed to \code{fun}.}
  \item{fun}{function.}
  \item{\dots}{additional parameters to be passed to \code{fun}.}

  }

\author{ A function borrowed from Kevin Sheppard's Matlab garch toolbox
  as implemented by Alexios Ghalanos in his \pkg{rgarch} package }
