\name{funBasics}

\alias{funBasics}


\title{Demo Function Addon}


\description{
        
     A collection of functions written as examples and listed
     in the demo file "funBasics".
     
     The functions are:
     
     \tabular{rll}{
         1  \tab    \code{print.dates}  
            \tab    S3 print mehtod for objects of class "dates", \cr 
         2  \tab    \code{print.chron}   
            \tab    S3 print mehtod for objects of class "chron", \cr 
         3  \tab    \code{seq.chron}    
            \tab    SPlus like "seq.dates" function, \cr
            
         4  \tab    \code{dssd}   
            \tab    Density using smoothing spline ANOVA model, \cr              
         5	\tab	\code{pssd}
        	\tab	Probability using smoothing spline ANOVA model, \cr
       	 6  \tab    \code{qssd}     
            \tab    Quantiles using smoothing spline ANOVA model, \cr
         7  \tab    \code{rssd}    
            \tab    Random Deviates using smoothing spline ANOVA model, \cr 
         8	\tab	\code{ssdFit}
        	\tab	Parameter Fit using smoothing spline ANOVA model, \cr
        	
         9  \tab    \code{bootMean}  
            \tab    Bootstrapped mean to get Confidence Intervals, \cr 
            
        10 	\tab	\code{data.csv, csvImport}
        	\tab	Imports data from comma delimited CSV files, \cr
        11 	\tab	\code{forecastsImport}
        	\tab	Downloads Monthly Market Data from "www.forecasts.org", \cr
        	
        12 	\tab	\code{is.weekday}
        	\tab	Test if a date is a weekday day or not, \cr\
        13 	\tab	\code{is.weekend}
        	\tab	Test if a date is a weekend day or not, \cr
        14 	\tab	\code{is.bizday}
        	\tab	Test if a date is a business day or not, \cr
        15 	\tab	\code{holiday.ZURICH}
        	\tab	Holiday Calendar for Zurich in Switzerland, \cr
        16 	\tab	\code{summary.timeSeries}
        	\tab	S3 Summary method for objects of class "timeDate". }   
                        
}


\examples{
## no examples ...
}


\keyword{programming}

