\name{FunctionAddon}

\alias{FunctionAddon}


\title{Demo Function Addon}


\description{
        
	A collection and description of functions written as 
	examples and listed in the demo file "funBasics".
	\cr
	
	The functions are:
     
	\tabular{ll}{
	\code{print.dates} \tab S3 print mehtod for objects of class "dates", \cr 
	\code{print.chron} \tab S3 print mehtod for objects of class "chron", \cr 
	\code{seq.chron} \tab SPlus like "seq.dates" function, \cr
            
	\code{dssd} \tab Density using smoothing spline ANOVA model, \cr              
	\code{pssd} \tab Probability using smoothing spline ANOVA model, \cr
	\code{qssd} \tab Quantiles using smoothing spline ANOVA model, \cr
	\code{rssd} \tab Random Deviates using smoothing spline ANOVA model, \cr 
	\code{ssdFit} \tab Parameter Fit using smoothing spline ANOVA model, \cr
        	
	\code{bootMean} \tab Bootstrapped mean to get Confidence Intervals, \cr 
            
	\code{data.csv, csvImport} \tab Imports data from comma delimited CSV files, \cr
	\code{forecastsImport} \tab Downloads Monthly Market Data from "www.forecasts.org", \cr
        	
	\code{is.weekday} \tab Test if a date is a weekday day or not, \cr\
	\code{is.weekend} \tab Test if a date is a weekend day or not, \cr
	\code{is.bizday} \tab Test if a date is a business day or not, \cr
	\code{holiday.ZURICH} \tab Holiday Calendar for Zurich in Switzerland, \cr
	\code{summary.timeSeries} \tab S3 Summary method for objects of class "timeDate". }   
                        
}


\author{

	Diethelm Wuertz for the Rmetrics \R-port.
	
}


\examples{
## funBasics -
   # The examples can be found in 
   # the demo file "funBasics".
}


\keyword{programming}

