\name{gldMode}

\alias{gldMode}

\title{Generalized Lambda Distribution Mode}

\description{

  Computes the mode of the generalized lambda distribution.

}

\usage{
gldMode(lambda1 = 0, lambda2 = -1, lambda3 = -1/8, lambda4 = -1/8)
}


\arguments{

  \item{lambda1, lambda2, lambda3, lambda4}{
    are numeric values where
    \code{lambda1} is the location parameter,
    \code{lambda2} is the location parameter,
    \code{lambda3} is the first shape parameter, and
    \code{lambda4} is the second shape parameter.
  }
}

\author{
  Implemented by Diethelm Wuertz
}
