% Part of the Rmetrics project, http://R-metrics.org/
% Copyright 2005-2011 Rmetrics Core Team
% Distributed under GPL 2 or later

\name{fBasics-deprecated}
%------ NOTE:  ../R/fBasics-deprecated.R   must be synchronized with this!
\alias{fBasics-deprecated}
%% \alias{dstable}
%% \alias{pstable}
%% \alias{qstable}
%% \alias{rstable}
%% \alias{stableMode}
\title{Deprecated Functions in Package fBasics}
\description{
  These functions are provided for compatibility with older versions of
  the package only, and may be defunct as soon as of the next release.

  There are none currently.
  \code{\link[stabledist]{dstable}} etc now are defunct, as they have
  been available from \pkg{stabledist} since early 2011.
%% dstable(x, alpha, beta, gamma = 1, delta = 0, pm = 0)
%% pstable(q, alpha, beta, gamma = 1, delta = 0, pm = 0)
%% qstable(p, alpha, beta, gamma = 1, delta = 0, pm = 0)
%% rstable(n, alpha, beta, gamma = 1, delta = 0, pm = 0)
%% stableMode(alpha, beta)
}
%% \usage{
%% }
%% \arguments{
%% }
\seealso{
  \code{\link{Deprecated}}, \code{\link{Defunct}}
}
\keyword{misc}
