\name{tsHessian}

\alias{tsHessian}

\title{Two sided approximated Hessian}

\description{
  Computes two sided (TS) approximated Hessian.
}

\usage{
tsHessian(x, fun, ...)
}

\arguments{

  \item{x}{arugment to be passed to function}
  \item{fun}{function}
  \item{\dots}{additional paramters to be passed to function}

  }

\author{ A function borrowed from Kevin Sheppard's Matlab garch toolbox
  as implemented by Alexios Ghalanos in his \pkg{rgarch} package }
