\name{FunctionAddon}

\alias{FunctionAddon}


\title{Demo Function Addon}


\description{
        
    A collection and description of functions written as 
    examples and listed in the demo file "funBasics".
    \cr
    
    The functions are:
     
    \tabular{ll}{            
    \code{bootMean} \tab Bootstrapped mean to get Confidence Intervals, \cr 
            
    \code{data.csv, csvImport} \tab Imports data from comma delimited CSV files, \cr
    \code{forecastsImport} \tab Downloads Monthly Market Data from "www.forecasts.org". }   
                        
}


\author{

    Diethelm Wuertz for the Rmetrics \R-port.
    
}


\examples{
## SOURCE("fBasics.A0-SPlusCompatibility")
## SOURCE("fBasics.Z1-BasicsTools")

## funBasics -
   # The examples can be found in 
   # the demo file "funBasics".
}


\keyword{programming}

